WIP Generating code to read the entire list
