--- pkg/gnmi-client-gen-v2/template/gnmi_client.go.tpl
@@ -1,33 +1,0 @@
-/*
-* SPDX-FileCopyrightText: 2022-present Intel Corporation
-*
-* SPDX-License-Identifier: Apache-2.0
-*/
-
-// Generated via gnmi-gen.go, do NOT edit
-
-package api
-
-import (
-    "context"
-    "fmt"
-    "github.com/onosproject/config-models/pkg/gnmi-client-gen/gnmi_utils"
-    "github.com/openconfig/gnmi/proto/gnmi"
-    "google.golang.org/grpc"
-    "google.golang.org/grpc/codes"
-    "google.golang.org/grpc/status"
-    "reflect"
-    "time"
-)
-
-type GnmiClient struct {
-    client gnmi.GNMIClient
-}
-
-func New{{ .PluginName }}GnmiClient(conn *grpc.ClientConn) *GnmiClient {
-    gnmi_client := gnmi.NewGNMIClient(conn)
-    return &GnmiClient{client: gnmi_client}
-}
-
-{{/* NOTE we always assume that the models are container in a root (Device) entry */ -}}
-{{ template "_entry.go.tpl" .Entry }}
