--- pkg/gnmi-client-gen-v2/testdata/results/basic-list.txt
@@ -1,79 +1,0 @@
-/*
-* SPDX-FileCopyrightText: 2022-present Intel Corporation
-*
-* SPDX-License-Identifier: Apache-2.0
-*/
-
-// Generated via gnmi-gen.go, do NOT edit
-
-package api
-
-import (
-    "context"
-    "fmt"
-    "github.com/onosproject/config-models/pkg/gnmi-client-gen/gnmi_utils"
-    "github.com/openconfig/gnmi/proto/gnmi"
-    "google.golang.org/grpc"
-    "google.golang.org/grpc/codes"
-    "google.golang.org/grpc/status"
-    "reflect"
-    "time"
-)
-
-type GnmiClient struct {
-    client gnmi.GNMIClient
-}
-
-func NewTestGnmiClient(conn *grpc.ClientConn) *GnmiClient {
-    gnmi_client := gnmi.NewGNMIClient(conn)
-    return &GnmiClient{client: gnmi_client}
-}
-
-func (c *GnmiClient) Get_List1(ctx context.Context, target string, list1_list1leaf2 uint16,
-) (*Test_List1, error) {
-	gnmiCtx, cancel := context.WithTimeout(ctx, 10*time.Second)
-	defer cancel()
-
-	path := []*gnmi.Path{
-		{
-			Elem: []*gnmi.PathElem{
-				{
-					Name: "list1",
-					Key: map[string]string{
-						"list1_list1leaf2": fmt.Sprint(list1_list1leaf2),
-					},
-				},
-			},
-			Target: target,
-		},
-	}
-
-	req := &gnmi.GetRequest{
-		Encoding: gnmi.Encoding_JSON,
-		Path:     path,
-	}
-	res, err := c.client.Get(gnmiCtx, req)
-
-	if err != nil {
-		return nil, err
-	}
-
-	val, err := gnmi_utils.GetResponseUpdate(res)
-
-	if err != nil {
-		return nil, err
-	}
-
-	json := val.GetJsonVal()
-	st := Device{}
-	Unmarshal(json, &st)
-
-	if reflect.ValueOf(st.List1).Kind() == reflect.Ptr && reflect.ValueOf(st.List1).IsNil() {
-		return nil, status.Error(codes.NotFound, "List1-not-found")
-	}
-	if res, ok := st.List1[list1_list1leaf2]; ok {
-		return res, nil
-	}
-
-	return nil, status.Error(codes.NotFound, "List1-not-found")
-}
