+++ pkg/gnmi-client-gen/generator.go
@@ -6,387 +6,387 @@
 
 package gnmi_client_gen
 
+//{{- /*gotype: github.com/openconfig/goyang/pkg/yang.Entry -}}
 import (
-	"errors"
 	"fmt"
 	t "github.com/onosproject/config-models/pkg/gnmi-client-gen/template"
-	"github.com/onosproject/onos-lib-go/pkg/logging"
 	"github.com/openconfig/goyang/pkg/yang"
 	"github.com/openconfig/ygot/genutil"
 	"golang.org/x/text/cases"
 	"golang.org/x/text/language"
 	"google.golang.org/grpc/codes"
 	"google.golang.org/grpc/status"
-	"os"
-	"sort"
+	"io"
 	"strings"
 	"text/template"
 )
 
 const templateFile = "gnmi_client.go.tpl"
 
-var log = logging.GetLogger("gnmi-client-gen")
-
-type GnmiEndpoints struct {
-	LeavesEndpoints    []LeavesEndpoint
-	ContainerEndpoints []ContainerEndpoint
-	ListEndpoints      []ListEndpoint
-	PluginName         string
+type templateData struct {
+	PluginName string
+	Entry      *yang.Entry
 }
 
-// LeavesEndpoint contains the information needed to generate the gNMI client for
-// elements which are at the end of the tree (leaves) and thus are simple types.
-type LeavesEndpoint struct {
-	Method            string // GET or SET (NOTE: do we need an Enum?)
-	MethodName        string // Get, Update and Delete
-	ModelName         string
-	Path              []string // A string representing the gNMI path (/ separated)
-	GoType            string
-	GoReturnType      string
-	GoEmptyReturnType string
+func isContainer(e *yang.Entry) bool {
+	return e.IsContainer()
 }
 
-type ContainerEndpoint struct {
-	ModelName       string
-	ModelPath       string
-	ParentModelPath string
-	Method          string   // GET or SET (NOTE: do we need an Enum?)
-	MethodName      string   // Get, Update and Delete + itemName
-	Path            []string // A list of strings representing the gNMI path
+func isList(e *yang.Entry) bool {
+	return e.IsList()
 }
 
-type ListKey struct {
-	Type      string   // the generated Key type (can be a Go type, eg: string)
-	Keys      []Key    // a list of keys that we need to set in the path
-	ParentKey *ListKey // a reference to a parent key, this is needed in case of nested lists
-	ModelName string   // represent the model in the path, to simplify recursion
-}
-type Key struct {
-	Name string
-	Type string
-	Ptr  bool // whether this is pointer or not
-}
-
-type ListEndpoint struct {
-	ContainerEndpoint
-	Key              ListKey
-	ParentPath       []string // A list of strings representing the gNMI path to the parent model
-	PluralMethodName string   // Used for the methods that apply to the entire list
-}
-
-func BuildGnmiStruct(debug bool, pluginName string, entry *yang.Entry, parentPath []string, parentKey *ListKey) (*GnmiEndpoints, error) {
-
-	if debug {
-		log.SetLevel(logging.DebugLevel)
+// takes tuple of parameters and returns a map
+// it's used to pass multiple values to a template
+func dict(values ...interface{}) (map[string]interface{}, error) {
+	if len(values)%2 != 0 {
+		return nil, status.Errorf(codes.InvalidArgument, "invalid dict call")
 	}
-
-	g := &GnmiEndpoints{
-		LeavesEndpoints:    []LeavesEndpoint{},
-		ContainerEndpoints: []ContainerEndpoint{},
-		ListEndpoints:      []ListEndpoint{},
-		PluginName:         pluginName,
-	}
-
-	for _, item := range entry.Dir {
-		itemPath := append(parentPath, item.Name)
-		log.Debugw("item", "path", itemPath)
-
-		if item.IsLeaf() || item.IsLeafList() {
-			eps, err := generateGnmiEndpointsForLeaf(item, itemPath)
-			if err != nil {
-				return nil, err
-			}
-			g.LeavesEndpoints = append(g.LeavesEndpoints, eps...)
-		} else if item.Kind == yang.ChoiceEntry {
-			// FIXME add support
-			log.Warn("item-is-choice-entry")
-		} else if item.IsContainer() {
-			_g, err := BuildGnmiStruct(debug, pluginName, item, itemPath, parentKey)
-			if err != nil {
-				return nil, err
-			}
-			g.LeavesEndpoints = append(g.LeavesEndpoints, _g.LeavesEndpoints...)
-			g.ContainerEndpoints = append(g.ContainerEndpoints, _g.ContainerEndpoints...)
-			g.ListEndpoints = append(g.ListEndpoints, _g.ListEndpoints...)
-
-			// generate endpoints for container
-			eps, err := generateGnmiEndpointsForContainer(item, itemPath)
-			if err != nil {
-				return nil, err
-			}
-			g.ContainerEndpoints = append(g.ContainerEndpoints, eps...)
-		} else if item.IsList() {
-			// generate the key for this list
-			key, err := GetListKey(item)
-			if err != nil {
-				if s, ok := status.FromError(err); ok && s.Code() == codes.Unimplemented {
-					log.Warnw("skipping-generation-for-list", "item", item.Name, "err", err.Error())
-					continue
-				} else {
-					return nil, err
-				}
-			}
-
-			// if there is parent key add it to key for this item
-			if parentKey != nil {
-				key.ParentKey = parentKey
-			}
-
-			// recurse down the spec
-			_g, err := BuildGnmiStruct(debug, pluginName, item, itemPath, &key)
-			if err != nil {
-				return nil, err
-			}
-			g.LeavesEndpoints = append(g.LeavesEndpoints, _g.LeavesEndpoints...)
-			g.ContainerEndpoints = append(g.ContainerEndpoints, _g.ContainerEndpoints...)
-			g.ListEndpoints = append(g.ListEndpoints, _g.ListEndpoints...)
-
-			// generate endpoints for lists
-			eps, err := generateGnmiEndpointsForLists(item, itemPath, &key)
-			if err != nil {
-				return nil, err
-			}
-			g.ListEndpoints = append(g.ListEndpoints, eps...)
-		} else {
-			return nil, fmt.Errorf("unhandled item %v.Type %v", item.Name, item.Type)
+	dict := make(map[string]interface{}, len(values)/2)
+	for i := 0; i < len(values); i += 2 {
+		key, ok := values[i].(string)
+		if !ok {
+			return nil, status.Errorf(codes.InvalidArgument, "dict keys must be strings")
 		}
+		dict[key] = values[i+1]
 	}
-	ep, err := sortByName(*g)
-	return &ep, err
+	return dict, nil
 }
 
-const gnmiGet = "get"
-const gnmiUpdate = "update"
-const gnmiDelete = "delete"
+// remove all chars that are not valid in variables/methods names
+func sanitize(s string) string {
+	return strings.ReplaceAll(s, "-", "")
+}
 
-// NOTE do we need a "create" or an update without an ID ==  a create?
-var methods = []string{gnmiGet, gnmiDelete, gnmiUpdate}
-
-func generateGnmiEndpointsForLeaf(item *yang.Entry, path []string) ([]LeavesEndpoint, error) {
-	eps := []LeavesEndpoint{}
-	itemName := PathToCamelCaseName(path)
-
+func capitalize(s string) string {
 	caser := cases.Title(language.English)
-	for _, m := range methods {
-		epName := caser.String(m)
+	return caser.String(s)
+}
 
-		t, err := yangTypeToGoType(item)
+func hasParent(entry *yang.Entry) bool {
+	return entry.Parent != nil
+}
+
+// returns the correct format for an existing value
+func goReturnVal(val *yang.Entry) (string, error) {
+	switch val.Type.Kind {
+	case yang.Yint8:
+		return "int8(val.GetIntVal())", nil
+	case yang.Yint16:
+		return "int16(val.GetIntVal())", nil
+	case yang.Yint32:
+		return "int32(val.GetIntVal())", nil
+	case yang.Yint64:
+		return "int64(val.GetIntVal())", nil
+	case yang.Yuint8:
+		return "uint8(val.GetUintVal())", nil
+	case yang.Yuint16:
+		return "uint16(val.GetUintVal())", nil
+	case yang.Yuint32:
+		return "uint32(val.GetUintVal())", nil
+	case yang.Yuint64:
+		return "uint64(val.GetUintVal())", nil
+	case yang.Ybool, yang.Yempty:
+		return "val.GetBoolVal()", nil
+	case yang.Ystring, yang.Yunion:
+		return "val.GetStringVal()", nil
+	case yang.Ydecimal64:
+		return "float64(val.GetFloatVal())", nil
+	case yang.Ybinary:
+		return "val.GetBytesVal()", nil
+	case yang.Yenum, yang.Yidentityref:
+		return "int64(val.GetIntVal())", nil
+	case yang.Yleafref:
+		t, err := findLeafRefType(val.Type.Path, val)
 		if err != nil {
-			if s, ok := status.FromError(err); ok && s.Code() == codes.Unimplemented {
-				log.Warnw("skipping-generation-for-leaf", "item", item.Name, "err", err.Error())
-				continue
-			} else {
-				return nil, err
-			}
+			return "", err
 		}
-		ep := LeavesEndpoint{
-			Method:            m,
-			ModelName:         itemName,
-			MethodName:        fmt.Sprintf("%s_%s", epName, itemName),
-			Path:              path,
-			GoType:            t,
-			GoReturnType:      yangTypeToGoReturnVal(item.Type.Kind),
-			GoEmptyReturnType: yangTypeToGoEmptyReturnVal(item.Type.Kind),
+
+		if t == "float64" {
+			t = "decimal64"
 		}
-		eps = append(eps, ep)
+
+		yt, ok := yang.TypeKindFromName[t]
+		if !ok {
+			return "", status.Errorf(codes.Internal, "type %s is not a valid yang kind", t)
+		}
+		return goReturnVal(&yang.Entry{Type: &yang.YangType{Kind: yt}})
 	}
-	log.Debugw("generating-methods-for-leaf-item",
-		"name", itemName, "path", path, "endpoints", eps)
-	return eps, nil
+
+	return "", status.Errorf(codes.Unimplemented, "%T type is not supported yet", val.Type.Kind)
 }
 
-func generateGnmiEndpointsForContainer(item *yang.Entry, path []string) ([]ContainerEndpoint, error) {
-	// NOTE YGOT model names are generated here: https://github.com/openconfig/ygot/blob/6f722a0cce2a47949294afa0c3f23b080d51e501/ygen/goelements.go#L193
-	eps := []ContainerEndpoint{}
-	// itemName is the name of the Model as generated by YGOT (without module prefix)
-	itemName := PathToYgotModelName(path)
-	// pathInModel is where this particular container is nested in the YGOT device
-	pathInModel := PathToYgotModelPath(path)
-	// parentModelPath is the path to the parent container
-	parentModelPath := PathToYgotModelPath(path[:len(path)-1])
-
-	caser := cases.Title(language.English)
-	for _, m := range methods {
-
-		epName := caser.String(m)
-
-		ep := ContainerEndpoint{
-			Method:          m,
-			MethodName:      fmt.Sprintf("%s_%s", epName, itemName),
-			Path:            path,
-			ModelName:       fmt.Sprintf("%s", item.Annotation["structname"]),
-			ModelPath:       pathInModel,
-			ParentModelPath: parentModelPath,
+// returns the correct format to return in case of error
+func goEmptyReturnVal(val *yang.Entry) (string, error) {
+	switch val.Type.Kind {
+	case yang.Yint8,
+		yang.Yint16,
+		yang.Yint32,
+		yang.Yint64,
+		yang.Yuint8,
+		yang.Yuint16,
+		yang.Yuint32,
+		yang.Yuint64,
+		yang.Ydecimal64,
+		yang.Yenum, yang.Yidentityref:
+		return "0", nil
+	case yang.Ybool, yang.Yempty:
+		return "false", nil
+	case yang.Ystring:
+		return "\"\"", nil
+	case yang.Ybinary:
+		return "nil", nil
+	case yang.Yleafref:
+		t, err := findLeafRefType(val.Type.Path, val)
+		if err != nil {
+			return "", err
 		}
-		eps = append(eps, ep)
-	}
-	log.Debugw("generating-methods-for-container-item",
-		"name", itemName, "path", path, "endpoints", eps)
-	return eps, nil
-}
 
-// I assume that a list always contains Yang containers, we can't have a list of strings
-func generateGnmiEndpointsForLists(item *yang.Entry, path []string, key *ListKey) ([]ListEndpoint, error) {
-
-	eps := []ListEndpoint{}
-
-	// itemName is the name of the Model as generated by YGOT (without module prefix)
-	itemName := PathToYgotModelName(path)
-
-	// pathInModel is where this particular container is nested in the YGOT device
-	pathInModel := PathToYgotModelPath(path)
-	// parentModelPath is the path to the parent container
-	parentModelPath := PathToYgotModelPath(path[:len(path)-1])
-
-	caser := cases.Title(language.English)
-	for _, m := range methods {
-
-		epName := caser.String(m)
-
-		ep := ListEndpoint{
-			ContainerEndpoint: ContainerEndpoint{
-				Method:          m,
-				MethodName:      fmt.Sprintf("%s_%s", epName, itemName),
-				Path:            path,
-				ModelName:       fmt.Sprintf("%s", item.Annotation["structname"]),
-				ModelPath:       pathInModel,
-				ParentModelPath: parentModelPath,
-			},
-			PluralMethodName: fmt.Sprintf("%s_%s_List", epName, itemName),
-			ParentPath:       path[:len(path)-1],
-			Key:              *key,
+		if t == "float64" {
+			t = "decimal64"
 		}
-		eps = append(eps, ep)
+
+		yt, ok := yang.TypeKindFromName[t]
+		if !ok {
+			return "", status.Errorf(codes.Internal, "type %s is not a valid yang kind", t)
+		}
+		return goEmptyReturnVal(&yang.Entry{Type: &yang.YangType{Kind: yt}})
+	case yang.Yunion:
+		return "\"\"", nil
 	}
-	log.Debugw("generating-methods-for-list-items",
-		"name", itemName, "path", path, "endpoints", eps)
-	return eps, nil
+
+	return "", status.Errorf(codes.Unimplemented, "%T type is not supported yet", val.Type.Kind)
 }
 
-// sortByName will order the lists based on the MethodName
-// a possible future improvement is to order them by ModelName while retaining the order Get, Update, Delete
-// but the main reason to have this is to make it easier to compare diffs in PRs for an easier review process
-func sortByName(endpoints GnmiEndpoints) (GnmiEndpoints, error) {
-	sort.Sort(leavesByMethodName(endpoints.LeavesEndpoints))
-	sort.Sort(containersByMethodName(endpoints.ContainerEndpoints))
-	sort.Sort(listsByMethodName(endpoints.ListEndpoints))
-	return endpoints, nil
+func goType(entry *yang.Entry) (string, error) {
+	// NOTE inspired by https://github.com/openconfig/ygot/blob/master/ytypes/util_types.go#L353
+	switch entry.Type.Kind {
+	case yang.Yint8:
+		return "int8", nil
+	case yang.Yint16:
+		return "int16", nil
+	case yang.Yint32:
+		return "int32", nil
+	case yang.Yint64:
+		return "int64", nil
+	case yang.Yuint8:
+		return "uint8", nil
+	case yang.Yuint16:
+		return "uint16", nil
+	case yang.Yuint32:
+		return "uint32", nil
+	case yang.Yuint64:
+		return "uint64", nil
+	case yang.Ybool, yang.Yempty:
+		return "bool", nil
+	case yang.Ystring:
+		return "string", nil
+	case yang.Ydecimal64:
+		return "float64", nil
+	case yang.Ybinary:
+		return "[]byte", nil
+	case yang.Yenum:
+		return "int64", nil
+	case yang.Yidentityref:
+		// FIXME this is not enough, eg:
+		// SYSLOG_FACILITY is generated via YGOT as `type E_OpenconfigSystemLogging_SYSLOG_FACILITY int64`
+		// NOTE that PrefixedName returns :SYSLOG_FACILITY
+		//return entry.Type.IdentityBase.PrefixedName()
+		return "", status.Error(codes.Unimplemented, "yang.Yidentityref type is not supported yet")
+	case yang.Yleafref:
+		v, err := findLeafRefType(entry.Type.Path, entry)
+		return v, err
+	case yang.Yunion:
+		return "string", nil
+	}
+	// not ideal, but for now we'll take it
+	return "", status.Error(codes.Unimplemented, fmt.Sprintf("%s type is not supported yet (entry: %s)", entry.Type.Kind.String(), entry.Name))
 }
 
-func ApplyTemplate(epList *GnmiEndpoints, outPath string) error {
+// extracts the structname from a yang entry
+func structName(entry *yang.Entry, forList bool) (string, error) {
+	name, ok := entry.Annotation["structname"]
+	if !ok {
+		return "", status.Errorf(codes.NotFound, "structname not found in annotations")
+	}
+	if entry.IsList() && forList {
+		keys, err := listKeys(entry)
+		if err != nil {
+			return "", err
+		}
+		// if the entry is a list and we are targeting the container
+		// NOTE we still need to support composite keys
+		return fmt.Sprintf("map[%s]*%s", keys[0].Gotype, name), nil
+	}
+	return fmt.Sprintf("*%s", name), nil
+}
 
-	var funcs template.FuncMap = map[string]interface{}{
-		"quote": func(value interface{}) string {
-			return fmt.Sprintf("\"%s\"", value)
-		},
-		"replace": func(search, replace string, value interface{}) string {
-			return strings.ReplaceAll(fmt.Sprint(value), search, replace)
-		},
-		"sub": func(a int, b int) int {
-			return a - b
-		},
-		"lower": strings.ToLower,
-		"toName": func(s string) string {
-			return genutil.EntryCamelCaseName(&yang.Entry{Name: s})
-		},
-		"hasParentKey": func(key ListKey) bool {
-			return key.ParentKey != nil
-		},
-		// takes tuple of parameters and returns a map
-		// it's used to pass multiple values to a template
-		"dict": func(values ...interface{}) (map[string]interface{}, error) {
-			if len(values)%2 != 0 {
-				return nil, errors.New("invalid dict call")
+// returns true if the entry is the root of the tree
+func isRoot(entry *yang.Entry) bool {
+	if entry.Annotation != nil {
+		if v, ok := entry.Annotation["isFakeRoot"]; ok {
+			return v == true
+		}
+	}
+	return false
+}
+
+// return the path to a model from the root
+func devicePath(entry *yang.Entry, forParent bool) (string, error) {
+
+	name := genutil.EntryCamelCaseName(entry)
+	if entry.IsList() {
+		var keyName string
+		// if the item is a list, we need to return the element in the list based on the key
+		keys := strings.Split(entry.Key, " ")
+		if len(keys) > 1 {
+			keyName = "key"
+		} else {
+			keyName = entry.Key
+		}
+		if !forParent {
+			name = sanitize(fmt.Sprintf("%s[%s]", name, fmt.Sprintf("%s_%s", entry.Name, keyName)))
+		}
+	}
+
+	if entry.Parent != nil && !isRoot(entry.Parent) {
+		// forParent applies only for the last level,
+		// in a structure like item1 -> item2 -> item3
+		// we return:
+		// - forParent = false => Item1.Item2.Item3
+		// - forParent = true => Item1.Item2
+		parentName, err := devicePath(entry.Parent, false)
+		if err != nil {
+			return "", err
+		}
+		if !forParent || entry.IsList() {
+			return fmt.Sprintf("%s.%s", parentName, name), nil
+		}
+		return parentName, nil
+	}
+
+	return name, nil
+}
+
+// extract the list keys from a yang entry
+// and returns the tuples
+type YangKey struct {
+	// Name represents the variable name in the method arguments,
+	// it's prefixed with the entry name to avoid duplicates, eg: application_id and device_id
+	Name string
+	// Key is the name of the key in the Yang path
+	Key string
+	// Gotype is the Go type that corresponds to a key, used to build the method parameters list
+	Gotype string
+}
+
+func listKeys(entry *yang.Entry) ([]YangKey, error) {
+	list := []YangKey{}
+	if !entry.IsList() {
+		// if this is not a list, return an empty list
+		return list, nil
+	}
+	for _, k := range strings.Split(entry.Key, " ") {
+		// find the child entry the key refers to get the type
+		if e, ok := entry.Dir[k]; ok {
+			// the only error case is for unsupported types, ignore for now
+			if t, err := goType(e); err == nil {
+				list = append(list, YangKey{
+					Name:   fmt.Sprintf("%s_%s", entry.Name, k),
+					Key:    k,
+					Gotype: t,
+				})
 			}
-			dict := make(map[string]interface{}, len(values)/2)
-			for i := 0; i < len(values); i += 2 {
-				key, ok := values[i].(string)
-				if !ok {
-					return nil, errors.New("dict keys must be strings")
-				}
-				dict[key] = values[i+1]
-			}
-			return dict, nil
-		},
+
+		}
+	}
+	return list, nil
+}
+
+func Generate(pluginName string, entry *yang.Entry, output io.Writer) error {
+
+	if entry == nil {
+		return fmt.Errorf("entry-cannot-be-nil")
+	}
+
+	templateFunctions := template.FuncMap{
+		"dict":             dict,
+		"isContainer":      isContainer,
+		"isList":           isList,
+		"sanitize":         sanitize,
+		"capitalize":       capitalize,
+		"hasParent":        hasParent,
+		"goReturnVal":      goReturnVal,
+		"goEmptyReturnVal": goEmptyReturnVal,
+		"goType":           goType,
+		"structName":       structName,
+		"devicePath":       devicePath,
+		"listKeys":         listKeys,
+		"isRoot":           isRoot,
 	}
 
 	t, err := template.New(templateFile).
-		Funcs(funcs).
+		Funcs(templateFunctions).
 		ParseFS(t.GnmiGenTemplate, "*.go.tpl")
 	if err != nil {
 		return err
 	}
 
-	file, err := os.Create(outPath)
-	if err != nil {
-		return err
-	}
-	defer file.Close()
-
-	return t.Execute(file, epList)
+	return t.Execute(output, templateData{
+		PluginName: pluginName,
+		Entry:      entry,
+	})
 }
 
-// helpers to the sortByName method
-func less(s []interface{}, i, j int) bool {
-	ok1 := false
-	ok2 := false
+func findLeafRefType(path string, entry *yang.Entry) (string, error) {
 
-	// cast to LeavesEndpoint
-	leaf1, ok1 := s[i].(LeavesEndpoint)
-	leaf2, ok2 := s[j].(LeavesEndpoint)
-	if ok1 && ok2 {
-		return leaf1.MethodName < leaf2.MethodName
+	var downwardPath = []string{}
+	var cur_entry = entry
+
+	// if the path is absolute (eg: /t1:cont1a/t1:list2a/t1:name) go back to the root and then descend
+	if path[0:1] == "/" {
+		dp := strings.Split(path[1:], "/")
+
+		// remove the prefix from the pieces
+		for _, p := range dp {
+			// a piece of the path can be ta:cont1a or cont1a
+			k := strings.Split(p, ":")
+			if len(k) > 1 {
+				downwardPath = append(downwardPath, k[1])
+			} else {
+				downwardPath = append(downwardPath, p)
+			}
+		}
+		// go back till the root
+		for {
+			if cur_entry.Parent != nil {
+				cur_entry = cur_entry.Parent
+			} else {
+				break
+			}
+		}
+
+	} else {
+		// identify how many levels we have to go up the tree
+		lb := strings.Count(path, "../")
+
+		// identify the path to take once we have gone up the tree
+		downwardPath = strings.Split(strings.ReplaceAll(path, "../", ""), "/")
+
+		// this is the entry we are moving to
+
+		for i := 0; i < lb; i++ {
+			// we're going up the tree till it's needed
+			cur_entry = cur_entry.Parent
+		}
+
+	}
+	for _, k := range downwardPath {
+		// and then descending to the leafref path
+		cur_entry = cur_entry.Dir[k]
 	}
 
-	cont1, ok1 := s[i].(ContainerEndpoint)
-	cont2, ok2 := s[j].(ContainerEndpoint)
-	if ok1 && ok2 {
-		return cont1.MethodName < cont2.MethodName
-	}
+	// not simply convert the type
+	return goType(cur_entry)
 
-	list1, ok1 := s[i].(ListEndpoint)
-	list2, ok2 := s[j].(ListEndpoint)
-	if ok1 && ok2 {
-		return list1.MethodName < list2.MethodName
-	}
-	panic(fmt.Sprintf("type-%T-not-supported-for-ordering", s[i]))
-}
-
-type leavesByMethodName []LeavesEndpoint
-
-func (s leavesByMethodName) Len() int      { return len(s) }
-func (s leavesByMethodName) Swap(i, j int) { s[i], s[j] = s[j], s[i] }
-func (s leavesByMethodName) Less(i, j int) bool {
-	// we have to manually convert to a list of interfaces
-	b := make([]interface{}, len(s))
-	for i := range s {
-		b[i] = s[i]
-	}
-	return less(b, i, j)
-}
-
-type containersByMethodName []ContainerEndpoint
-
-func (s containersByMethodName) Len() int      { return len(s) }
-func (s containersByMethodName) Swap(i, j int) { s[i], s[j] = s[j], s[i] }
-func (s containersByMethodName) Less(i, j int) bool {
-	// we have to manually convert to a list of interfaces
-	b := make([]interface{}, len(s))
-	for i := range s {
-		b[i] = s[i]
-	}
-	return less(b, i, j)
-}
-
-type listsByMethodName []ListEndpoint
-
-func (s listsByMethodName) Len() int      { return len(s) }
-func (s listsByMethodName) Swap(i, j int) { s[i], s[j] = s[j], s[i] }
-func (s listsByMethodName) Less(i, j int) bool {
-	// we have to manually convert to a list of interfaces
-	b := make([]interface{}, len(s))
-	for i := range s {
-		b[i] = s[i]
-	}
-	return less(b, i, j)
 }
