+++ pkg/gnmi-client-gen/sampleSchemas/aether-2.1.x.json
