+++ pkg/gnmi-client-gen/sampleSchemas/testdevice-1.json
