+++ pkg/gnmi-client-gen/template/_entry.go.tpl
