+++ pkg/gnmi-client-gen/template/_entry_name.go.tpl
