+++ pkg/gnmi-client-gen/testdata/results.go
